+++ ProgramFiles/CMakeLists.txt
@@ -17,6 +17,10 @@ add_executable (Client
     "NSS/src/clients.cpp"
   )
 
+add_executable (Server 
+    "NSS/server/release.cpp"
+  )
+
 
 if (CMAKE_VERSION VERSION_GREATER 3.12)
   set_property(TARGET NSS PROPERTY CXX_STANDARD 20)
